requirements.txt
Django
djangorestframework
pyJWT
bcrypt